feat: Auth 추가, token paylode값 추가
- token이 있고 변경하려는 user와 요청하는 user가 같을 경우에만 Update, Delete가 가능하도록 변경함
- 해당 기능 test 추가
- token에 iss, sub, 추가

+++ services/UserServices.py
@@ -4,7 +4,11 @@ from schema.user import UserCreate, UserUpdate, validate_password
 from models.user import User
 from datetime import datetime
 from passlib.context import CryptContext
+from fastapi.security import OAuth2PasswordBearer
+from jose import jwt
+from core.config import settings
 
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
 
@@ -52,10 +56,13 @@ class UserService:
             )
         )
 
-    def update(self, id: int, user_body: UserUpdate) -> User:
+    def update(self, id: int, user_body: UserUpdate, req_user: User) -> User:
         user_exist = self.get(id)
         if not user_exist:
             raise HTTPException(status_code=404, detail="사용자를 찾을 수 없습니다.")
+
+        if req_user.id != user_exist.id:
+            raise HTTPException(status_code=401, detail="권한이 없습니다.")
 
         if validate_password(user_body.password):
             user_body.password = pwd_context.hash(user_body.password)
@@ -67,8 +74,17 @@ class UserService:
             id, User(nickname=user_body.nickname, password=user_body.password)
         )
 
-    def delete(self, id: int) -> None:
+    def delete(self, id: int, req_user: User) -> None:
         user_exist = self.get(id)
         if not user_exist:
             raise HTTPException(status_code=404, detail="사용자를 찾을 수 없습니다.")
+
+        if req_user.id != user_exist.id:
+            raise HTTPException(status_code=401, detail="권한이 없습니다.")
+
         return self.userRepository.delete(id)
+
+    def get_current_user(self, token: str = Depends(oauth2_scheme)):
+        decoded_token = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
+        req_user = self.get(int(decoded_token["sub"]))
+        return req_user
